Separate built-in profiles, cli, and cmd modules
This commit moves the built-in profiles for HIV1b and HCV1a out of the
`cli` into their own modules under `alignmentprofile/builtin`. After
this commit, the `cli` module only imports the alignment profile
struct so it can accept it as a parameter, and the `cmd` module only
imports the built-in profiles so that it can pass them through to the
`cli`.
